requirements.txt
feedparser>=6.0
requests>=2.31
schedule>=1.2
python-dotenv>=1.0
google-genai>=1.0
notion-client>=2.2
